% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chymo_high.R
\docType{data}
\name{Chymo_high}
\alias{Chymo_high}
\title{Product concentration of 101 observed time}
\format{A data frame with 101 rows and 2 variables:
\describe{
\item{time}{observed time, no unit}
\item{product}{product concentration, no unit}
}}
\usage{
Chymo_high
}
\description{
An aritficial data set containing the product concentration
and it's observed time based on high enzyme concentration and
low sustrate concentration
}
\keyword{datasets}
